Moved autosplitter management to autosplitter.cpp
