+++ src/autosplitter.h
@@ -1,0 +1,10 @@
+#ifndef AUTOSPLITTER_H
+#define AUTOSPLITTER_H
+
+extern string autoSplittersDirectory;
+
+void checkDirectories();
+void chooseAutoSplitter();
+void runAutoSplitter();
+
+#endif
